WIP Refactoring / cleaning the compose and ECS services related settings
