+++ ecs_composex/alarms/alarms_stack.py
@@ -20,14 +20,14 @@ from ecs_composex.alarms.alarms_elbv2 import (
 )
 from ecs_composex.alarms.alarms_params import RES_KEY
 from ecs_composex.common import build_template, setup_logging
-from ecs_composex.common.compose_resources import (
+from ecs_composex.common.stacks import ComposeXStack
+from ecs_composex.compose.x_resources import (
     XResource,
     set_lookup_resources,
     set_new_resources,
     set_resources,
     set_use_resources,
 )
-from ecs_composex.common.stacks import ComposeXStack
 from ecs_composex.resources_import import import_record_properties
 
 LOG = setup_logging()
