+++ ecs_composex/codeguru_profiler/codeguru_profiler_params.py
@@ -1,12 +1,12 @@
-﻿#   -*- coding: utf-8 -*-
+#   -*- coding: utf-8 -*-
 #  SPDX-License-Identifier: MPL-2.0
 #  Copyright 2020-2021 John Mille <john@compose-x.io>
 
 from os import path
 
 from ecs_composex.common import NONALPHANUM
-from ecs_composex.common.compose_resources import Parameter
 from ecs_composex.common.ecs_composex import X_KEY
+from ecs_composex.compose.x_resources import Parameter
 
 MOD_KEY = path.basename(path.dirname(path.abspath(__file__)))
 RES_KEY = f"{X_KEY}{MOD_KEY}"
