+++ ecs_composex/cognito_userpool/cognito_params.py
@@ -5,8 +5,8 @@
 import re
 from os import path
 
-from ecs_composex.common.compose_resources import Parameter
 from ecs_composex.common.ecs_composex import X_KEY
+from ecs_composex.compose.x_resources import Parameter
 
 MOD_KEY = path.basename(path.dirname(path.abspath(__file__)))
 RES_KEY = f"{X_KEY}{MOD_KEY}"
