+++ ecs_composex/cognito_userpool/cognito_userpool_stack.py
@@ -22,14 +22,14 @@ from ecs_composex.cognito_userpool.cognito_params import (
     USERPOOL_NAME,
 )
 from ecs_composex.common import build_template, setup_logging
-from ecs_composex.common.compose_resources import (
+from ecs_composex.common.stacks import ComposeXStack
+from ecs_composex.compose.x_resources import (
     XResource,
     set_lookup_resources,
     set_new_resources,
     set_resources,
     set_use_resources,
 )
-from ecs_composex.common.stacks import ComposeXStack
 
 LOG = setup_logging()
