+++ ecs_composex/compose/__init__.py
@@ -1,0 +1,3 @@
+#   -*- coding: utf-8 -*-
+#  SPDX-License-Identifier: MPL-2.0
+#  Copyright 2020-2021 John Mille <john@compose-x.io>
