+++ ecs_composex/compose/compose_networks.py
@@ -1,4 +1,9 @@
-﻿#  -*- coding: utf-8 -*-
+#  -*- coding: utf-8 -*-
+
+#   -*- coding: utf-8 -*-
+#  SPDX-License-Identifier: MPL-2.0
+#  Copyright 2020-2021 John Mille <john@compose-x.io>
+
 # SPDX-License-Identifier: MPL-2.0
 # Copyright 2020-2021 John Mille <john@compose-x.io>
