+++ ecs_composex/compose/compose_secrets.py
@@ -1,4 +1,9 @@
-﻿#  -*- coding: utf-8 -*-
+#  -*- coding: utf-8 -*-
+
+#   -*- coding: utf-8 -*-
+#  SPDX-License-Identifier: MPL-2.0
+#  Copyright 2020-2021 John Mille <john@compose-x.io>
+
 # SPDX-License-Identifier: MPL-2.0
 # Copyright 2020-2021 John Mille <john@compose-x.io>
 
@@ -6,32 +11,18 @@
 Represent a service from the docker-compose services
 """
 
-import re
 from copy import deepcopy
 
+from compose_x_common.aws.secrets_manager import get_secret_name_from_arn
 from compose_x_common.compose_x_common import keyisset
 from troposphere import AWS_ACCOUNT_ID, AWS_PARTITION, AWS_REGION, FindInMap, Sub
 from troposphere.ecs import Secret as EcsSecret
 
 from ecs_composex.common import LOG, NONALPHANUM
-from ecs_composex.ecs.ecs_params import EXEC_ROLE_T, TASK_ROLE_T
+from ecs_composex.ecs.ecs_params import EXEC_ROLE_T
 from ecs_composex.kms.kms_params import KMS_KEY_ARN_RE
 from ecs_composex.secrets.secrets_aws import lookup_secret_config
 from ecs_composex.secrets.secrets_params import RES_KEY, XRES_KEY
-
-
-def get_name_from_arn(secret_arn):
-    secret_re = re.compile(
-        r"(?:^arn:aws(?:-[a-z]+)?:secretsmanager:[\w-]+:[0-9]{12}:secret:)([\S]+)(?:-[A-Za-z0-9]{1,6})$"
-    )
-    if not secret_re.match(secret_arn):
-        raise ValueError(
-            "The secret ARN is invalid",
-            secret_arn,
-            "No name cound be found from it via",
-            r"(?:^arn:aws(?:-[a-z]+)?:secretsmanager:[\w-]+:[0-9]{12}:secret:)([\S]+)(?:-[A-Za-z0-9]{1,6})$",
-        )
-    return secret_re.match(secret_arn).groups()[0]
 
 
 def match_secrets_services_config(service, s_secret, secrets):
@@ -168,25 +159,22 @@ class ComposeSecret(object):
             self.define_names_from_lookup(settings.session)
 
         self.define_links()
-        self.validate_links()
         if self.mapping:
             settings.secrets_mappings.update({self.logical_name: self.mapping})
             self.add_json_keys()
 
     def add_json_keys(self):
         """
-        Method to add secrets definitions based on JSON secret keys
+        Add secrets definitions based on JSON secret keys
         """
         if not keyisset(self.json_keys_key, self.definition[self.x_key]):
             return
         required_keys = ["SecretKey"]
         allowed_keys = ["SecretKey", "VarName", "Transform"]
         unfiltered_secrets = self.definition[self.x_key][self.json_keys_key]
-        LOG.debug(f"UN-FILTERED SECRETS {unfiltered_secrets}")
         filtered_secrets = [
             dict(y) for y in set(tuple(x.items()) for x in unfiltered_secrets)
         ]
-        LOG.debug(f"FILTERED SECRETS {filtered_secrets}")
         for secret_key in filtered_secrets:
             if not all(key in allowed_keys for key in secret_key):
                 raise KeyError(
@@ -233,7 +221,7 @@ class ComposeSecret(object):
 
     def define_names_from_import(self):
         """
-        Method to define the names from docker-compose file content
+        Define the names from docker-compose file content
         """
         if not keyisset(self.map_name_name, self.definition[self.x_key]):
             raise KeyError(
@@ -241,7 +229,7 @@ class ComposeSecret(object):
             )
         name_input = self.definition[self.x_key][self.map_name_name]
         if name_input.startswith("arn:"):
-            self.aws_name = get_name_from_arn(
+            self.aws_name = get_secret_name_from_arn(
                 self.definition[self.x_key][self.map_name_name]
             )
             self.mapping = {
@@ -282,13 +270,12 @@ class ComposeSecret(object):
     def define_names_from_lookup(self, session):
         """
         Method to Lookup the secret based on its tags.
-        :return:
         """
         lookup_info = self.definition[self.x_key]["Lookup"]
         if keyisset("Name", self.definition[self.x_key]):
             lookup_info["Name"] = self.definition[self.x_key]["Name"]
         secret_config = lookup_secret_config(self.logical_name, lookup_info, session)
-        self.aws_name = get_name_from_arn(secret_config[self.logical_name])
+        self.aws_name = get_secret_name_from_arn(secret_config[self.logical_name])
         self.arn = secret_config[self.logical_name]
         self.iam_arn = secret_config[self.logical_name]
         if keyisset("KmsKeyId", secret_config) and not secret_config[
@@ -313,18 +300,8 @@ class ComposeSecret(object):
         self.ecs_secret = [EcsSecret(Name=self.name, ValueFrom=self.arn)]
 
     def define_links(self):
+        """
+        Defines which IAM role to assign the secrets access policy to. Defaults to exec role
+        """
         if keyisset(self.links_key, self.definition[self.x_key]):
             self.links = self.definition[self.x_key][self.links_key]
-
-    def validate_links(self):
-        if not isinstance(self.links, list):
-            raise TypeError("LinksTo must be of type", list, "Got", type(self.links))
-        for link in self.links:
-            if link not in [EXEC_ROLE_T, TASK_ROLE_T]:
-                raise ValueError(
-                    "Links in LinksTo can only be one of",
-                    EXEC_ROLE_T,
-                    TASK_ROLE_T,
-                    "Got",
-                    link,
-                )
