+++ ecs_composex/dynamodb/dynamodb_stack.py
@@ -13,14 +13,14 @@ from troposphere import GetAtt, Ref
 from troposphere.dynamodb import Table as CfnTable
 
 from ecs_composex.common import build_template, setup_logging
-from ecs_composex.common.compose_resources import (
+from ecs_composex.common.stacks import ComposeXStack
+from ecs_composex.compose.x_resources import (
     XResource,
     set_lookup_resources,
     set_new_resources,
     set_resources,
     set_use_resources,
 )
-from ecs_composex.common.stacks import ComposeXStack
 from ecs_composex.dynamodb.dynamodb_params import (
     MAPPINGS_KEY,
     MOD_KEY,
