+++ ecs_composex/ecs/ecs_params.py
@@ -14,7 +14,7 @@ You can change the names *values* so you like so long as you keep it [a-zA-Z0-9]
 from troposphere import ImportValue, Ref, Select, Split, Sub
 
 from ecs_composex.common.cfn_params import ROOT_STACK_NAME_T, Parameter
-from ecs_composex.common.ecs_composex import CFN_EXPORT_DELIMITER as delim
+from ecs_composex.common.ecs_composex import CFN_EXPORT_DELIMITER as DELIM
 from ecs_composex.vpc.vpc_params import SG_ID_TYPE
 
 CLUSTER_T = "EcsCluster"
@@ -185,6 +185,6 @@ def get_import_service_group_id(remote_service_name):
     """
     return ImportValue(
         Sub(
-            f"${{{ROOT_STACK_NAME_T}}}{delim}{remote_service_name}{delim}{SERVICE_GROUP_ID_T}"
+            f"${{{ROOT_STACK_NAME_T}}}{DELIM}{remote_service_name}{DELIM}{SERVICE_GROUP_ID_T}"
         )
     )
