+++ ecs_composex/ecs_composex.py
@@ -71,6 +71,7 @@ SUPPORTED_X_MODULE_NAMES = [
     "codeguru_profiler",
     "ssm_parameter",
     "opensearch",
+    "neptune",
 ]
 
 SUPPORTED_X_MODULES = [f"{X_KEY}{mod_name}" for mod_name in SUPPORTED_X_MODULE_NAMES]
@@ -84,7 +85,16 @@ EXCLUDED_X_KEYS = [
     f"{X_KEY}cluster",
     f"{X_KEY}dashboards",
 ]
-TCP_MODES = ["rds", "appmesh", "elbv2", "docdb", "elasticache", "efs", "opensearch"]
+TCP_MODES = [
+    "rds",
+    "appmesh",
+    "elbv2",
+    "docdb",
+    "elasticache",
+    "efs",
+    "opensearch",
+    "neptune",
+]
 TCP_SERVICES = [f"{X_KEY}{mode}" for mode in TCP_MODES]
