+++ ecs_composex/elasticache/elasticache_stack.py
@@ -1,4 +1,4 @@
-﻿#  -*- coding: utf-8 -*-
+#  -*- coding: utf-8 -*-
 # SPDX-License-Identifier: MPL-2.0
 # Copyright 2020-2021 John Mille <john@compose-x.io>
 
@@ -12,14 +12,14 @@ from compose_x_common.compose_x_common import keyisset
 from troposphere import GetAtt, Ref, Sub
 from troposphere.ssm import Parameter as SSMParameter
 
-from ecs_composex.common.compose_resources import (
+from ecs_composex.common.stacks import ComposeXStack
+from ecs_composex.compose.x_resources import (
     XResource,
     set_lookup_resources,
     set_new_resources,
     set_resources,
     set_use_resources,
 )
-from ecs_composex.common.stacks import ComposeXStack
 from ecs_composex.elasticache.elasticache_ecs import create_lookup_mappings
 from ecs_composex.elasticache.elasticache_params import (
     CLUSTER_CONFIG,
