+++ ecs_composex/neptune/neptune_stack.py
@@ -6,14 +6,12 @@
 AWS DocumentDB entrypoint for ECS ComposeX
 """
 
-from botocore.exceptions import ClientError
-from compose_x_common.aws.rds import RDS_DB_CLUSTER_ARN_RE
+from compose_x_common.aws.neptune import NEPTUNE_DB_CLUSTER_ARN_RE
 from compose_x_common.compose_x_common import attributes_to_mapping, keyisset
 from troposphere import GetAtt, Ref
-from troposphere.docdb import DBCluster as CfnDBCluster
+from troposphere.neptune import DBCluster as CfnDBCluster
 
 from ecs_composex.common import setup_logging
-from ecs_composex.common.aws import find_aws_resource_arn_from_tags_api
 from ecs_composex.common.stacks import ComposeXStack
 from ecs_composex.compose.x_resources import (
     XResource,
@@ -22,27 +20,31 @@ from ecs_composex.compose.x_resources import (
     set_resources,
     set_use_resources,
 )
-from ecs_composex.docdb.docdb_params import (
-    DOCDB_NAME,
-    DOCDB_PORT,
-    DOCDB_SECRET,
-    DOCDB_SG,
-    MAPPINGS_KEY,
-    MOD_KEY,
-    RES_KEY,
-)
 from ecs_composex.docdb.docdb_template import (
     create_docdb_template,
     init_doc_db_template,
 )
-from ecs_composex.rds_resources_settings import lookup_rds_resource, lookup_rds_secret
+from ecs_composex.iam.import_sam_policies import get_access_types
+from ecs_composex.neptune.neptune_params import (
+    DB_CLUSTER_ARN,
+    DB_CLUSTER_NAME,
+    DB_CLUSTER_RESOURCES_ARN,
+    DB_ENDPOINT,
+    DB_PORT,
+    DB_READ_ENDPOINT,
+    DB_SG,
+    MAPPINGS_KEY,
+    MOD_KEY,
+    RES_KEY,
+)
+from ecs_composex.rds_resources_settings import lookup_rds_resource
 from ecs_composex.vpc.vpc_params import STORAGE_SUBNETS
 
 LOG = setup_logging()
 
 
 def get_db_cluster_config(db, account_id, resource_id):
-    client = db.lookup_session.client("docdb")
+    client = db.lookup_session.client("neptune")
     try:
         db_config_r = client.describe_db_clusters(
             DBClusterIdentifier=db.arn,
@@ -50,7 +52,7 @@ def get_db_cluster_config(db, account_id, resource_id):
                 {
                     "Name": "engine",
                     "Values": [
-                        "docdb",
+                        "neptune",
                     ],
                 },
             ],
@@ -68,19 +70,24 @@ def get_db_cluster_config(db, account_id, resource_id):
         ]
 
     attributes_mappings = {
-        DOCDB_PORT.title: "Port",
+        DB_PORT.title: "Port",
         "VpcSecurityGroupId": "VpcSecurityGroups::0::VpcSecurityGroupId",
+        DB_CLUSTER_ARN.title: DB_CLUSTER_ARN.title,
     }
     config = attributes_to_mapping(db_cluster, attributes_mappings)
+    config[
+        DB_CLUSTER_RESOURCES_ARN.title
+    ] = f"{config[DB_CLUSTER_ARN.title].replace('rds', 'neptune-db', 1)}/*"
     return config
 
 
-class DocDb(XResource):
+class NeptuneDBCluster(XResource):
     """
     Class to manage DocDB
     """
 
     subnets_param = STORAGE_SUBNETS
+    policies_scaffolds = get_access_types(MOD_KEY)
 
     def __init__(self, name, definition, module_name, settings, mapping_key=None):
         """
@@ -90,7 +97,6 @@ class DocDb(XResource):
         :param dict definition:
         :param ecs_composex.common.settings.ComposeXSettings settings:
         """
-        self.db_secret = None
         self.db_sg = None
         self.db_subnets_group = None
         super().__init__(
@@ -103,53 +109,45 @@ class DocDb(XResource):
         Method to init the DocDB output attributes
         """
         self.output_properties = {
-            DOCDB_NAME: (self.logical_name, self.cfn_resource, Ref, None),
-            DOCDB_PORT: (
-                f"{self.logical_name}{DOCDB_PORT.return_value}",
+            DB_CLUSTER_NAME: (self.logical_name, self.cfn_resource, Ref, None),
+            DB_CLUSTER_ARN: (
+                f"{self.logical_name}{DB_CLUSTER_ARN.title}",
                 self.cfn_resource,
-                GetAtt,
-                DOCDB_PORT.return_value,
-            ),
-            DOCDB_SECRET: (
-                self.db_secret.title,
-                self.db_secret,
                 Ref,
                 None,
             ),
-            DOCDB_SG: (
-                self.db_sg.title,
+            DB_CLUSTER_RESOURCES_ARN: (
+                f"{self.logical_name}{DB_CLUSTER_RESOURCES_ARN.title}",
+                self.cfn_resource,
+                Ref,
+                None,
+            ),
+            DB_ENDPOINT: (
+                f"{self.logical_name}{DB_ENDPOINT.return_value}",
+                self.cfn_resource,
+                GetAtt,
+                DB_ENDPOINT.return_value,
+            ),
+            DB_READ_ENDPOINT: (
+                f"{self.logical_name}{DB_READ_ENDPOINT.return_value}",
+                self.cfn_resource,
+                GetAtt,
+                DB_READ_ENDPOINT.return_value,
+            ),
+            DB_PORT: (
+                f"{self.logical_name}{DB_PORT.return_value}",
+                self.cfn_resource,
+                GetAtt,
+                DB_PORT.return_value,
+            ),
+            DB_SG: (
+                f"{self.logical_name}{DB_SG.return_value}",
                 self.db_sg,
                 GetAtt,
-                DOCDB_SG.return_value,
+                DB_SG.return_value,
+                "VpcSecurityGroupId",
             ),
         }
-
-    def lookup_secret(self, secret_lookup):
-
-        if keyisset("Arn", secret_lookup):
-            client = self.lookup_session.client("secretsmanager")
-            try:
-                secret_arn = client.describe_secret(SecretId=secret_lookup["Arn"])[
-                    "ARN"
-                ]
-            except client.exceptions.ResourceNotFoundException:
-                LOG.error(
-                    f"{self.module_name}.{self.name} - Secret {secret_lookup['Arn']} not found"
-                )
-                raise
-            except ClientError as error:
-                LOG.error(error)
-                raise
-        elif keyisset("Tags", secret_lookup):
-            secret_arn = find_aws_resource_arn_from_tags_api(
-                self.lookup["secret"], self.lookup_session, "secretsmanager:secret"
-            )
-        else:
-            raise LookupError(
-                f"{self.module_name}.{self.name} - Failed to find the DB Secret"
-            )
-        if secret_arn:
-            self.mappings[DOCDB_SECRET.title] = secret_arn
 
     def lookup_resource(
         self,
@@ -179,7 +177,9 @@ class XStack(ComposeXStack):
     """
 
     def __init__(self, title, settings, **kwargs):
-        set_resources(settings, DocDb, RES_KEY, MOD_KEY, mapping_key=MAPPINGS_KEY)
+        set_resources(
+            settings, NeptuneDBCluster, RES_KEY, MOD_KEY, mapping_key=MAPPINGS_KEY
+        )
         x_resources = settings.compose_content[RES_KEY].values()
         new_resources = set_new_resources(x_resources, RES_KEY, True)
         lookup_resources = set_lookup_resources(x_resources, RES_KEY)
@@ -195,14 +195,11 @@ class XStack(ComposeXStack):
                 settings.mappings[RES_KEY] = {}
             for resource in lookup_resources:
                 resource.lookup_resource(
-                    RDS_DB_CLUSTER_ARN_RE,
+                    NEPTUNE_DB_CLUSTER_ARN_RE,
                     get_db_cluster_config,
                     CfnDBCluster.resource_type,
                     "rds:cluster",
-                    "cluster",
                 )
-                if keyisset("secret", resource.lookup):
-                    lookup_rds_secret(resource, resource.lookup["secret"])
                 settings.mappings[RES_KEY].update(
                     {resource.logical_name: resource.mappings}
                 )
