+++ ecs_composex/s3/s3_stack.py
@@ -6,9 +6,6 @@
 Module to control S3 stack
 """
 
-import json
-
-from boto3.session import Session
 from botocore.exceptions import ClientError
 from compose_x_common.aws import get_account_id
 from compose_x_common.aws.kms import (
@@ -23,14 +20,14 @@ from troposphere.s3 import Bucket as CfnBucket
 
 from ecs_composex.common import build_template, setup_logging
 from ecs_composex.common.aws import find_aws_resource_arn_from_tags_api
-from ecs_composex.common.compose_resources import (
+from ecs_composex.common.stacks import ComposeXStack
+from ecs_composex.compose.x_resources import (
     XResource,
     set_lookup_resources,
     set_new_resources,
     set_resources,
     set_use_resources,
 )
-from ecs_composex.common.stacks import ComposeXStack
 from ecs_composex.s3.s3_params import (
     CONTROL_CLOUD_ATTR_MAPPING,
     MAPPINGS_KEY,
