+++ ecs_composex/ssm_parameter/ssm_parameter_stack.py
@@ -16,14 +16,14 @@ from troposphere.ssm import Parameter as CfnSsmParameter
 from yaml import Loader
 
 from ecs_composex.common import LOG, add_outputs, build_template
-from ecs_composex.common.compose_resources import (
+from ecs_composex.common.stacks import ComposeXStack
+from ecs_composex.compose.x_resources import (
     XResource,
     set_lookup_resources,
     set_new_resources,
     set_resources,
     set_use_resources,
 )
-from ecs_composex.common.stacks import ComposeXStack
 from ecs_composex.iam.import_sam_policies import get_access_types
 from ecs_composex.resources_import import import_record_properties
 from ecs_composex.ssm_parameter.ssm_parameter_params import (
