+++ tests/pytests/test_secrets.py
@@ -1,8 +1,8 @@
-﻿#  -*- coding: utf-8 -*-
+#  -*- coding: utf-8 -*-
 # SPDX-License-Identifier: MPL-2.0
 # Copyright 2020-2021 John Mille<john@compose-x.io>
 
-from ecs_composex.secrets.compose_secrets import define_env_var_name
+from ecs_composex.compose.compose_secrets import define_env_var_name
 
 
 def test_normal_secrets():
